backend/requirements.txt
Flask==3.1.0
pymongo[srv]==4.10.1
bcrypt==4.2.1
flask-cors==5.0.1
python-dotenv==1.0.1
